pkg/whitelist/gen: set back oldestBrowserHistoryItemDate
The "last visit dates" in our test browser history files are static
and so limiting urls which only were accessed in the last N days will
eventually cut off all urls. We need to set this back to some date
which will allow all records in our test history files to be collected.