pkg/whitelist/gen: fallback to system roots if chain is empty
Perhaps a breaking change, but if there aren't roots provided
whitelist generation fails. We should just try to generate with the
system roots instead.

I ran into this on a docker image and firefox (or chrome) with only
one url in their history. I think that because only a leaf cert was
passed through that connections would fail.